refactor(settings): 设置页 IA 重构 + Windows 字体锐化
按独立 UX/IA 审计重构，保留一级 5 节框架，只迁卡片 + 节内分组。

- 拆「高级」杂物抽屉：日志/诊断 → 日志页(排障归一，折叠卡标题显当前级别)；clash API/终端速查 → 网络节；内核管理(回滚/重置出厂/完全卸载) 关于→高级运维区(破坏性运维不藏只读页)。高级现 = 备份 + 内核管理。
- general 启动/行为与隐私 拆为独立卡，与网络节卡片范式一致。
- network 减负：测速+订阅合并「更新与测速」(含更新走代理)；DNS/连接低频项归「高级 DNS」「高级流量」分组(默认展开，不隐藏控件)。
- UI 风格统一：设置行卡 divide-y / 动作区无分隔线 / 折叠组补 border-t 对齐节奏；卡头字重统一 semibold；五节描述随内容更新。
- 修 Windows 字体发虚：-webkit-font-smoothing:antialiased 限定 html.platform-darwin，Windows/Linux 回落系统默认恢复 ClearType。
- i18n zh-CN/en-US 同步；gate 全绿(lint/typecheck/jest 802/build)。纯前端 IA 迁移，真机(Mac/Win)待核对渲染与 Windows 字体。
