+++ src/renderer/components/settings/diagnostic-section.tsx
@@ -93,7 +93,7 @@ export function DiagnosticSection() {
   return (
     <div className="space-y-4">
       <div>
-        <h4 className="flex items-center gap-2 text-sm font-medium">
+        <h4 className="flex items-center gap-2 text-sm font-semibold">
           <Stethoscope className="h-3.5 w-3.5 text-muted-foreground" />
           {t('settings.advanced.diagnostics.title')}
         </h4>
